% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extentions.R
\name{create_nhl_db}
\alias{create_nhl_db}
\title{Create NHL database}
\usage{
create_nhl_db(db_folder = "db")
}
\arguments{
\item{db_folder}{Folder in which database will be created.}
}
\description{
IN DEVELOPMENT. Creates an SQLite database to hold nhlscrpr data. Need only run once per database instance.
}

