% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{current.games}
\alias{current.games}
\title{current.games}
\usage{
current.games(rdata.folder = "nhlr-data")
}
\arguments{
\item{rdata.folder}{Folder in which data from nhlscrapr is saved.}
}
\description{
Lists currently process games that are available for use with \code{\link{retrieve.game}}
}

