% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extentions.R
\name{retrieve_season}
\alias{retrieve_season}
\title{Retrieve season}
\usage{
retrieve_season(season, database = "db/nhl_db")
}
\arguments{
\item{season}{Character. Season to extract from database, e.g. "20162017."}

\item{database}{String. Location of database to read data from.}
}
\value{
data.frame containing data for season.
}
\description{
Retrieves a season of data from the database.
}

