% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlplot-rink-elements.R
\name{rink.plot}
\alias{rink.plot}
\title{Plot blank rink for use with points()}
\usage{
rink.plot(fresh = TRUE, thiscol = "gray85", ...)
}
\arguments{
\item{fresh}{Whether or not to start with rink.plot.blank() for fresh plot}

\item{thiscol}{Base color for goal crease, line segments, faceoff cirlces}

\item{...}{Args passed to rink.plot.}
}
\description{
Plot blank rink for use with points()
}

