% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlplot-goldsberry-plots.R
\name{run_dotplot}
\alias{run_dotplot}
\title{run_dotplot}
\usage{
run_dotplot(grand.data, plot.etype = c("SHOT", "GOAL"),
  plot.colors = c("red", "green"))
}
\arguments{
\item{grand.data}{A game (or season/team/whatever) dataframe.}

\item{plot.etype}{Vector indicating event types to be plotted.}

\item{plot.colors}{Vector giving color for each event type. Must be multiple of event type.}
}
\description{
Produces a simple dot plot based on either total shots or total goals for a game dataframe.
}

