% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extentions.R
\name{update_season_db}
\alias{update_season_db}
\title{Update Season in Database}
\usage{
update_season_db(grand.data)
}
\arguments{
\item{grand.data}{Grand.data events dataframe to be written to database.}
}
\value{
Returns list of gcodes that are held in updated database table for chosen season.
}
\description{
Renews database records for specified season in database.
}
\details{
This function first removes all data from current season and then re-writes all data to database from the grand.data file. Seasons are automatically detected from the grand.data file.
}

