% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlplot-goldsberry-plots.R
\name{run_dotplot}
\alias{run_dotplot}
\title{run_dotplot}
\usage{
run_dotplot(grand.data, type = "sg")
}
\arguments{
\item{grand.data}{A game (or season/team/whatever) dataframe}

\item{type}{"sg" for a combined shots and goals plot, "g" for a goals-only plot}
}
\description{
Produces a simple dot plot based on either total shots or total goals for a game dataframe.
}

